Increase subscription expiration time

+++ graph/update-subscriptions/index.js
@@ -65,21 +65,22 @@ module.exports = async function (context, myTimer) {
             continue;
         }
 
-        // Create a Date 1 day in the future
-        let now = new Date();
-        let newSubscriptionExpirationDateTime = new Date(now.setDate(now.getDate() + 1)).toISOString();
+        // Create a Date 2 days in the future
+        let expirationDateTime = new Date();
+        expirationDateTime.setDate(expirationDateTime.getDate() + 2);
+        let newSubscriptionExpirationDateTime = expirationDateTime.toISOString();
         splunk.logInfo(`[update-subscriptions] updating subscription ${JSON.stringify(subscription)}. New expiration: ${newSubscriptionExpirationDateTime}`);
         
         await graph.updateSubscriptionExpiration(subscription.subscriptionId, newSubscriptionExpirationDateTime)
             .catch((err) => {
-                if (err.statusCode = 404) {
+                if ((err.statusCode = 404) && (!err.message.includes("timed out"))) {
                     // Looks like this subscription was removed, so remove it from the blob container.
                     let msg = `[update-subscriptions] a subscription with subscription Id '${subscription.subscriptionId}' was not found from Graph. Removing the blob item...`
                     context.log.warn(msg);
                     splunk.logWarning(msg);
                     containerClient.deleteBlob(subscription.subscriptionId);
                 } else {
-                    let errorMsg = `[update-subscriptions] could not update subscription from Graph: ${subscriptions[i]}, error: ${err}`
+                    let errorMsg = `[update-subscriptions] could not update subscription from Graph: ${subscription.subscriptionId}, error: ${JSON.stringify(err)}`
                     context.log.error(errorMsg);
                     splunk.logError(errorMsg);
                     context.res = {
